# src/immich_memories/web/openapi.json
{
  "components": {
    "schemas": {
...
      "Choice": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "label"
        ],
        "title": "Choice",
        "type": "object"
      },
...
      "Decision": {
        "properties": {
          "action": {
            "enum": [
              "never_use",
              "clear",
              "forget"
            ],
            "title": "Action",
            "type": "string"
          },
          "level": {
            "default": "anyone",
            "enum": [
              "anyone",
              "family",
              "just-us"
            ],
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "Decision",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Hold": {
        "description": "`never_use`, `cleared:<level>` or None; `reasons` say what holds the picture, if anything.",
        "properties": {
          "can_clear": {
            "title": "Can Clear",
            "type": "boolean"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "decision",
          "reasons",
          "can_clear"
        ],
        "title": "Hold",
        "type": "object"
      },
...
      "LinkAnswer": {
        "properties": {
          "decision": {
            "anyOf": [
              {
                "enum": [
                  "confirmed",
                  "rejected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "decision"
        ],
        "title": "LinkAnswer",
        "type": "object"
      },
...
      "NewPerson": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "NewPerson",
        "type": "object"
      },
      "PersonAnswers": {
        "properties": {
          "links": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LinkAnswer"
            },
            "title": "Links",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "title": "PersonAnswers",
        "type": "object"
      },
...
      "Pool": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PoolItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "Pool",
        "type": "object"
      },
      "PoolItem": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "fate": {
            "title": "Fate",
            "type": "string"
          },
          "favourite": {
            "title": "Favourite",
            "type": "boolean"
          },
          "hold": {
            "$ref": "#/components/schemas/Hold"
          },
          "in_cut": {
            "title": "In Cut",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "photo",
              "video",
              "live"
            ],
            "title": "Kind",
            "type": "string"
          },
          "taken": {
            "title": "Taken",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "taken",
          "kind",
          "favourite",
          "in_cut",
          "fate",
          "hold"
        ],
        "title": "PoolItem",
        "type": "object"
      },
...
      "Relationship": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id"
        ],
        "title": "Relationship",
        "type": "object"
      },
...
      "Roster": {
        "properties": {
          "flags": {
            "items": {
              "$ref": "#/components/schemas/RosterFlag"
            },
            "title": "Flags",
            "type": "array"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/RosterPerson"
            },
            "title": "People",
            "type": "array"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/Choice"
            },
            "title": "Relationships",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "people",
          "flags",
          "relationships",
          "roles"
        ],
        "title": "Roster",
        "type": "object"
      },
      "RosterFlag": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "names": {
            "items": {
              "type": "string"
            },
            "title": "Names",
            "type": "array"
          },
          "person_ids": {
            "items": {
              "type": "string"
            },
            "title": "Person Ids",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "names",
          "person_ids",
          "message"
        ],
        "title": "RosterFlag",
        "type": "object"
      },
      "RosterLink": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "inferred": {
            "title": "Inferred",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "reverse_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverse Kind"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_name": {
            "title": "Target Name",
            "type": "string"
          },
          "via": {
            "title": "Via",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "target_id",
          "target_name",
          "confidence",
          "via",
          "inferred"
        ],
        "title": "RosterLink",
        "type": "object"
      },
      "RosterPerson": {
        "properties": {
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "counts_reliable": {
            "title": "Counts Reliable",
            "type": "boolean"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/RosterLink"
            },
            "title": "Links",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "person_id": {
            "title": "Person Id",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "person_id",
          "name",
          "birth_date",
          "tier",
          "count",
          "counts_reliable",
          "evidence",
          "links"
        ],
        "title": "RosterPerson",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Immich Memories",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pictures/{asset_id}/decision": {
      "post": {
        "description": "The owner's word on one picture, kept across runs: never use it, clear its hold, or forget.",
        "operationId": "decide_api_v1_pictures__asset_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hold"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "pool"
        ]
      }
    },
    "/api/v1/roster": {
      "get": {
        "description": "Everyone in the people file, inner circle first, with what needs curating.",
        "operationId": "roster_api_v1_roster_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roster"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Roster",
        "tags": [
          "people"
        ]
      },
      "post": {
        "description": "Add someone Immich has not tagged, or who is never on camera.",
        "operationId": "add_api_v1_roster_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPerson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterPerson"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/v1/roster/{person_id}": {
      "put": {
        "description": "Keep this person's role, notes and answers to the graph's guesses.",
        "operationId": "answer_api_v1_roster__person_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonAnswers"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterPerson"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer",
        "tags": [
          "people"
        ]
      }
    },
    "/api/v1/roster/{person_id}/relationships": {
      "delete": {
        "description": "Remove a relationship somebody recorded, and its reciprocal.",
        "operationId": "unrelate_api_v1_roster__person_id__relationships_delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Relationship"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterPerson"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unrelate",
        "tags": [
          "people"
        ]
      },
      "post": {
        "description": "Record one relationship; the file keeps its reciprocal.",
        "operationId": "relate_api_v1_roster__person_id__relationships_post",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Relationship"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterPerson"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relate",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/pool": {
      "get": {
        "description": "Every picture the cut saw, in capture order, with its fate and whatever holds it.",
        "operationId": "read_pool_api_v1_runs__run_id__pool_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 120,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pool"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Pool",
        "tags": [
          "pool"
        ]
      }
    },
...
  }
}